doc_processing_pipeline/docs_extractor_module/requirements.txt
fastapi
uvicorn
boto3
python-multipart

langchain-google-genai
langchain-core
python-dotenv
pandas
openpyxl